Extract DirectoryExists from utils:files
Bug: b/540567828
