+++ base/cvd/cuttlefish/common/libs/utils/files.cpp
@@ -63,6 +63,7 @@
 #include "cuttlefish/common/libs/utils/users.h"
 #include "cuttlefish/files/copy.h"
 #include "cuttlefish/files/directory_contents.h"
+#include "cuttlefish/files/directory_exists.h"
 #include "cuttlefish/files/file_device_id.h"
 #include "cuttlefish/files/file_exists.h"
 #include "cuttlefish/files/link_or_copy.h"
@@ -87,7 +88,7 @@ bool FileHasContent(const std::string& path) { return FileSize(path) > 0; }
 
 Result<void> LinkOrCopyDirectoryContentsRecursively(
     const std::string& source, const std::string& destination) {
-  CF_EXPECTF(IsDirectory(source), "Source '{}' is not a directory", source);
+  CF_EXPECTF(DirectoryExists(source), "Source '{}' is not a directory", source);
 
   CF_EXPECT(EnsureDirectoryExists(destination, 0755));
 
@@ -97,7 +98,7 @@ Result<void> LinkOrCopyDirectoryContentsRecursively(
     const std::string src_path = filepath;
     const std::string dst_path =
         destination + "/" + filepath.substr(source.size() + 1);
-    if (IsDirectory(src_path)) {
+    if (DirectoryExists(src_path)) {
       CF_EXPECT(EnsureDirectoryExists(dst_path));
       return {};
     }
@@ -112,7 +113,7 @@ Result<void> LinkOrCopyDirectoryContentsRecursively(
 
 Result<void> MoveDirectoryContents(const std::string& source,
                                    const std::string& destination) {
-  CF_EXPECTF(IsDirectory(source), "Source '{}' is not a directory", source);
+  CF_EXPECTF(DirectoryExists(source), "Source '{}' is not a directory", source);
   CF_EXPECT(EnsureDirectoryExists(destination));
 
   bool should_rename = CF_EXPECT(CanRename(source, destination));
@@ -139,17 +140,6 @@ Result<std::vector<std::string>> DirectoryContentsPaths(
     filename = fmt::format("{}/{}", path, filename);
   }
   return result;
-}
-
-bool DirectoryExists(const std::string& path, bool follow_symlinks) {
-  struct stat st{};
-  if ((follow_symlinks ? stat : lstat)(path.c_str(), &st) == -1) {
-    return false;
-  }
-  if ((st.st_mode & S_IFMT) != S_IFDIR) {
-    return false;
-  }
-  return true;
 }
 
 Result<void> EnsureDirectoryExists(const std::string& directory_path,
