+++ base/cvd/cuttlefish/host/commands/assemble_cvd/flags.cc
@@ -52,6 +52,7 @@
 #include "cuttlefish/common/libs/utils/host_info.h"
 #include "cuttlefish/common/libs/utils/known_paths.h"
 #include "cuttlefish/common/libs/utils/network.h"
+#include "cuttlefish/files/directory_exists.h"
 #include "cuttlefish/files/file_exists.h"
 #include "cuttlefish/host/commands/assemble_cvd/alloc.h"
 #include "cuttlefish/host/commands/assemble_cvd/assemble_cvd_flags.h"
