+++ base/cvd/cuttlefish/host/commands/cvd/cli/BUILD.bazel
@@ -79,6 +79,7 @@ cf_cc_library(
     deps = [
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/files:directory_contents",
+        "//cuttlefish/files:directory_exists",
         "//cuttlefish/host/commands/cvd/utils",
         "//cuttlefish/result",
         "@abseil-cpp//absl/log",
@@ -123,6 +124,7 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/common/libs/utils:users",
         "//cuttlefish/files:copy",
+        "//cuttlefish/files:directory_exists",
         "//cuttlefish/files:file_exists",
         "//cuttlefish/files:recursively_remove_directory",
         "//cuttlefish/flag_parser",
