+++ base/cvd/cuttlefish/host/commands/cvd/instances/instance_manager.cpp
@@ -30,6 +30,7 @@
 #include "fmt/format.h"
 
 #include "cuttlefish/common/libs/utils/files.h"
+#include "cuttlefish/files/directory_exists.h"
 #include "cuttlefish/files/recursively_remove_directory.h"
 #include "cuttlefish/host/commands/cvd/cli/commands/host_tool_target.h"
 #include "cuttlefish/host/commands/cvd/instances/local_instance.h"
