+++ base/cvd/cuttlefish/host/commands/cvd/instances/lock/BUILD.bazel
@@ -23,6 +23,7 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:contains",
         "//cuttlefish/common/libs/utils:environment",
         "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/files:directory_exists",
         "//cuttlefish/host/commands/cvd/utils:common",
         "//cuttlefish/posix:strerror",
         "//cuttlefish/result",
