+++ base/cvd/cuttlefish/host/commands/cvd/instances/lock/lock_file.cpp
@@ -32,6 +32,7 @@
 #include "cuttlefish/common/libs/fs/shared_fd.h"
 #include "cuttlefish/common/libs/utils/environment.h"
 #include "cuttlefish/common/libs/utils/files.h"
+#include "cuttlefish/files/directory_exists.h"
 #include "cuttlefish/result/result.h"
 
 namespace cuttlefish {
