+++ base/cvd/cuttlefish/host/commands/cvd/utils/BUILD.bazel
@@ -15,6 +15,7 @@ cf_cc_library(
     deps = [
         "//cuttlefish/common/libs/utils:contains",
         "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/files:directory_exists",
         "//cuttlefish/files:file_exists",
         "//cuttlefish/host/libs/config:config_utils",
         "//cuttlefish/result",
