+++ base/cvd/cuttlefish/host/commands/cvd/utils/common.cpp
@@ -25,6 +25,7 @@
 #include "absl/strings/strip.h"
 
 #include "cuttlefish/common/libs/utils/files.h"
+#include "cuttlefish/files/directory_exists.h"
 #include "cuttlefish/files/file_exists.h"
 #include "cuttlefish/host/libs/config/config_utils.h"
 #include "cuttlefish/result/result.h"
