+++ base/cvd/cuttlefish/host/commands/run_cvd/launch/BUILD.bazel
@@ -499,7 +499,7 @@ cf_cc_library(
     hdrs = ["tombstone_receiver.h"],
     deps = [
         "//cuttlefish/common/libs/fs",
-        "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/files:directory_exists",
         "//cuttlefish/host/libs/config:cuttlefish_config",
         "//cuttlefish/host/libs/config:known_paths",
         "//cuttlefish/host/libs/feature",
