+++ base/cvd/cuttlefish/host/commands/stop/BUILD.bazel
@@ -14,6 +14,7 @@ cf_cc_binary(
         "//cuttlefish/common/libs/utils:environment",
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/common/libs/utils:tee_logging",
+        "//cuttlefish/files:directory_exists",
         "//cuttlefish/files:recursively_remove_directory",
         "//cuttlefish/flag_parser",
         "//cuttlefish/host/libs/command_util",
